Add update via PATCH and ControllerAdvice to handle exceptions

+++ src/main/java/com/rest/repository/DiecastCarRepository.java
@@ -10,7 +10,7 @@ import java.util.Collection;
 import java.util.List;
 
 @Repository
-public interface DiecastCarRepository extends CrudRepository<DiecastCar, Integer> {
+public interface DiecastCarRepository extends CrudRepository<DiecastCar, Long> {
 
     //using JPQL to achieve abstraction and database independence
     @Query(value = "SELECT d FROM DiecastCar d WHERE d.brand IN :brands ")
